Update README.md
Online examination system is a user friendly system, which is very easy and convenient to use. The system is complete in the sense that it is operational and it is tested by entering data and getting the reports in proper order. But there is always a scope for improvement and enhancement. During the development of this ,coding standards are followed for easy maintainability and extensibility. 
The key concept is to minimize the amount of paper and convert all forms of documentation to digital form. It can observe with ease and accuracy in the computerized system. The user with minimum knowledge about computer can be able to operate the system easily. The system also produces brief result required by the management.